split post-comments up
